requirements.txt
google_api_python_client
numpy
protobuf
streamlit
streamlit_option_menu
streamlit_lottie
